convert timestamp to double, improves precision

+++ src/asterix/DataBlock.cpp
@@ -28,7 +28,7 @@
 
 extern bool gFiltering;
 
-DataBlock::DataBlock(Category *cat, unsigned long len, const unsigned char *data, unsigned long nTimestamp)
+DataBlock::DataBlock(Category *cat, unsigned long len, const unsigned char *data, double nTimestamp)
         : m_pCategory(cat), m_nLength(len), m_nTimestamp(nTimestamp), m_bFormatOK(false) {
     const unsigned char *m_pItemDataStart = data;
     long nUnparsed = len;
@@ -40,7 +40,7 @@ DataBlock::DataBlock(Category *cat, unsigned long len, const unsigned char *data
     }
 
     while (nUnparsed > 0) {
-        DataRecord *dr = new DataRecord(cat, counter++, nUnparsed, m_pItemDataStart, (unsigned long) nTimestamp);
+        DataRecord *dr = new DataRecord(cat, counter++, nUnparsed, m_pItemDataStart, nTimestamp);
 
         if (!dr) {
             Tracer::Error("Error DataBlock format.");
