+++ src/asterix/InputParser.h
@@ -38,10 +38,10 @@ public:
 
     ~InputParser() { if (m_pDefinition) delete m_pDefinition; }
 
-    AsterixData *parsePacket(const unsigned char *m_pBuffer, unsigned int m_nBufferSize, unsigned long nTimestamp = 0);
+    AsterixData *parsePacket(const unsigned char *m_pBuffer, unsigned int m_nBufferSize, double nTimestamp = 0.0);
 
     DataBlock *parse_next_data_block(const unsigned char *m_pData, unsigned int &m_nPos, unsigned int m_nBufferSize,
-                                     unsigned long nTimestamp, unsigned int &m_nDataLength);
+                                     double nTimestamp, unsigned int &m_nDataLength);
 
     std::string printDefinition();
